+++ src/test/java/ru/moysklad/remap_1_2/utils/TestAsserts.java
@@ -42,5 +42,6 @@ public interface TestAsserts {
         assertEquals(expected.getApartment(), real.getApartment());
         assertEquals(expected.getAddInfo(), real.getAddInfo());
         assertEquals(expected.getComment(), real.getComment());
+        assertEquals(expected.getFiasCode__ru(), real.getFiasCode__ru());
     }
 }